# backend/application/src/main/resources/META-INF/openapi.yaml
openapi: 3.1.0
info:
  title: Surplasse API
  version: 0.1.0
  summary: The Surplasse REST API, defined contract-first.
  description: |
    Surplasse lets independent restaurants run their own direct ordering
    channel. This contract is the single source of truth of the API: the
    backend implements Java interfaces generated from it, the frontends
    consume a generated TypeScript client.

    Version 0.1 covers the public catalog, ordering, payment, restaurateur
    identity flows, the authenticated operational-order list and the
    order-intake control. Other Dashboard reads remain drafts (`x-draft`
    blocks, excluded from generation and compatibility checks) until their
    authorization is implemented end to end.

    Conventions: paths are prefixed with `/v1`, money amounts are integers
    in cents with a `Cents` suffix, timestamps are ISO 8601 UTC, and every
    error is an RFC 9457 Problem Details document.
  contact:
    name: Surplasse
    url: https://surplasse.com
servers:
  - url: /
    description: Current API origin, injected by the environment at runtime.
...
paths:
...
  /v1/table-sessions:
    post:
      tags: [order]
      operationId: createTableSession
      summary: Open an anonymous table session from a scanned QR code
      description: |
        Exchanges the table code carried by a scanned QR
        (`{slug}.surplasse.com/?table={code}`) for an opaque anonymous
        session token, bound to the establishment and to that table. The
        token authorizes the ordering endpoints (`X-Table-Session` header)
        for the duration of a meal (2 hours, sliding while active). It
        carries no personal data. An unknown slug, an unknown or inactive
        table code, an establishment that is not active or an operational
        order-intake pause all yield a 404 without distinction.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TableSessionRequest"
      responses:
        "201":
          description: The anonymous session, bound to the establishment and the table.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TableSession"
              example:
                token: ts_9f2c4b7a1e6d8a3c5f0b2d4e6a8c1f3b
                establishmentId: 7c9e6679-7425-40de-944b-e07fc1f90ae7
                tableLabel: Table 4
                expiresAt: "2026-07-18T14:30:00Z"
        "400":
          description: The payload is syntactically invalid.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              example:
                type: https://surplasse.com/problems/validation-error
                title: Validation error
                status: 400
                detail: "tableCode must not be blank."
                instance: /v1/table-sessions
        "404":
          description: Unknown establishment or table code, or ordering closed.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              example:
                type: https://surplasse.com/problems/resource-not-found
                title: Resource not found
                status: 404
                detail: No active table matches this code.
                instance: /v1/table-sessions
  /v1/orders:
    get:
      tags: [order]
      operationId: listOrders
      summary: List the operational orders of an establishment
      description: |
        Cursor-paginated list of the operational orders of one
        establishment for the Dashboard, sorted by creation timestamp and
        identifier, both descending. Only `paid`, `accepted`, `preparing`
        and `ready` orders are returned. Pending payment and terminal orders
        are deliberately excluded from this phase.

        The opaque cursor stays stable while new orders arrive and is never
        interpreted by the client. The restaurateur session cookie and
        establishment membership are both required: an unknown establishment
        and an establishment outside the caller's scope yield the same 404.
      security:
        - restaurateurSession: []
      parameters:
        - name: establishmentId
          in: query
          required: true
          description: Establishment whose orders are listed.
          schema:
            type: string
            format: uuid
          example: 7c9e6679-7425-40de-944b-e07fc1f90ae7
        - name: cursor
          in: query
          required: false
          description: Opaque cursor returned by the previous page.
          schema:
            type: string
            minLength: 1
            maxLength: 512
          example: MjAyNi0wNy0xOFQxMjozNDo1NlonMGQ5ZThjN2I
        - name: limit
          in: query
          required: false
          description: Page size, between 1 and 100.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
          example: 50
      responses:
        "200":
          description: One page of orders, newest first.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderPage"
              example:
                items:
                  - id: 0d9e8c7b-6a5f-4e3d-8c2b-1a0f9e8d7c6b
                    displayNumber: "17"
                    status: paid
                    type: on_site
                    tableLabel: Table 4
                    lines:
                      - productId: d0e1f2a3-b4c5-46d7-88e9-f0a1b2c3d405
                        productName: Burger du Vieux-Port
                        unitPriceCents: 1600
                        quantity: 1
                        options: []
                        lineTotalCents: 1600
                    totalCents: 1600
                    currency: EUR
                    createdAt: "2026-07-18T12:34:56Z"
                nextCursor: djF8MjAyNi0wNy0xOFQxMjozNDo1NlowZDllOGM3Yg
                hasMore: true
        "400":
          description: The establishment identifier, limit or cursor is invalid.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              example:
                type: https://surplasse.com/problems/validation-error
                title: Validation error
                status: 400
                detail: The pagination cursor is malformed or incompatible with this request.
                instance: /v1/orders
        "401":
          description: The restaurateur session is missing or expired.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              example:
                type: https://surplasse.com/problems/session-expired
                title: Session expired
                status: 401
                detail: The restaurateur session is expired or missing.
                instance: /v1/orders
        "404":
          description: Unknown establishment, or establishment outside the caller's scope.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              example:
                type: https://surplasse.com/problems/resource-not-found
                title: Resource not found
                status: 404
                detail: No establishment matches this identifier.
                instance: /v1/orders
    post:
      tags: [order]
      operationId: createOrder
      summary: Create an order from a validated cart
      description: |
        Creates an order at status `pending_payment` from the cart validated
        by the customer. The cart itself never exists server side. The
        establishment and the table come exclusively from the anonymous
        table session (a customer can only order for the table whose QR was
        actually scanned). The server recomputes every amount from the
        catalog; amounts are never accepted from the client. The
        `Idempotency-Key` header guarantees that an unstable connection can
        never create the same order twice. Only `on_site` orders are
        accepted at this stage; `takeaway` is documented for the target and
        rejected with a business rule error until it opens. An order-intake
        pause rejects every new intention with `order-intake-paused`, while
        an exact replay still returns the order created before the pause.
      security:
        - tableSession: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderCreationRequest"
      responses:
        "201":
          description: The order was created, waiting for payment.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
              examples:
                order:
                  $ref: "#/components/examples/OrderExample"
        "400":
          description: The payload is syntactically invalid.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              example:
                type: https://surplasse.com/problems/validation-error
                title: Validation error
                status: 400
                detail: "lines must contain at least one element."
                instance: /v1/orders
        "401":
          description: The table session token is missing, unknown or expired.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              example:
                type: https://surplasse.com/problems/table-session-expired
                title: Table session expired
                status: 401
                detail: The table session token is expired or unknown.
                instance: /v1/orders
        "404":
          description: A product or option of the cart does not belong to the published menu.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              example:
                type: https://surplasse.com/problems/resource-not-found
                title: Resource not found
                status: 404
                detail: Product 6f4b9c2d-8a1e-4d7b-b3f5-0e9a8c7d6b5a is not on the menu.
                instance: /v1/orders
        "409":
          description: Ordering was paused, a product became unavailable, or the idempotency key was reused with a different payload.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                orderIntakePaused:
                  summary: New orders are paused
                  value:
                    type: https://surplasse.com/problems/order-intake-paused
                    title: Order intake paused
                    status: 409
                    detail: This establishment is not accepting new orders right now.
                    instance: /v1/orders
                productUnavailable:
                  summary: A product became unavailable
                  value:
                    type: https://surplasse.com/problems/product-unavailable
                    title: Product unavailable
                    status: 409
                    detail: Product 'Soupe de poisson' is marked unavailable on this menu.
                    instance: /v1/orders
        "422":
          description: A business rule rejected the cart (invalid option choices, takeaway not open).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              example:
                type: https://surplasse.com/problems/business-rule-violation
                title: Business rule violation
                status: 422
                detail: "Option group 'Cuisson' requires exactly one choice."
                instance: /v1/orders
  /v1/orders/{orderId}:
    get:
      tags: [order]
      operationId: getOrder
      summary: Get an order for its tracking page
      description: |
        Returns an order for the customer tracking page. Access is
        authorized by the order tracking token (a non-guessable capability
        returned at creation, carried by the tracking page URL): the
        customer can close the browser and come back. A wrong token yields
        a 404 without distinction.
      parameters:
        - $ref: "#/components/parameters/OrderId"
        - $ref: "#/components/parameters/TrackingToken"
      responses:
        "200":
          description: The order, with its current status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
              examples:
                order:
                  $ref: "#/components/examples/OrderExample"
        "400":
          description: The tracking token is missing or malformed.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              example:
                type: https://surplasse.com/problems/validation-error
                title: Validation error
                status: 400
                detail: "Query parameter 'trackingToken' is required."
                instance: /v1/orders/0d9e8c7b-6a5f-4e3d-8c2b-1a0f9e8d7c6b
        "404":
          description: Unknown order, or tracking token mismatch.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              example:
                type: https://surplasse.com/problems/resource-not-found
                title: Resource not found
                status: 404
                detail: No order matches this tracking token.
                instance: /v1/orders/0d9e8c7b-6a5f-4e3d-8c2b-1a0f9e8d7c6b
  /v1/orders/{orderId}/status:
    patch:
      tags: [order]
      operationId: updateOrderStatus
      summary: Advance an order through the operational workflow
      description: |
        Advances one order on behalf of an authenticated restaurateur. The
        target must be the next state allowed by the order state machine:
        `paid` to `accepted`, `accepted` to `preparing`, `preparing` to
        `ready`, then `ready` to `served` for an on-site order or `picked_up`
        for a takeaway order. Repeating the state already reached is
        idempotent and returns the current result without emitting another
        event. Refunds are deliberately excluded: they require a payment
        operation and cannot be represented by a status-only update.

        The restaurateur session cookie is required. Unknown orders and
        orders outside the caller's establishment scope yield the same 404.
      security:
        - restaurateurSession: []
      parameters:
        - $ref: "#/components/parameters/OrderId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: The order status after the idempotent update.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderStatusResult"
              example:
                id: 0d9e8c7b-6a5f-4e3d-8c2b-1a0f9e8d7c6b
                status: accepted
        "400":
          description: The order identifier or payload is syntactically invalid.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              example:
                type: https://surplasse.com/problems/validation-error
                title: Validation error
                status: 400
                detail: "status must be one of: accepted, preparing, ready, served, picked_up."
                instance: /v1/orders/0d9e8c7b-6a5f-4e3d-8c2b-1a0f9e8d7c6b/status
        "401":
          description: The restaurateur session is missing or expired.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              example:
                type: https://surplasse.com/problems/session-expired
                title: Session expired
                status: 401
                detail: The restaurateur session is expired or missing.
                instance: /v1/orders/0d9e8c7b-6a5f-4e3d-8c2b-1a0f9e8d7c6b/status
        "404":
          description: Unknown order, or order outside the caller's establishment scope.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              example:
                type: https://surplasse.com/problems/resource-not-found
                title: Resource not found
                status: 404
                detail: No order matches this identifier.
                instance: /v1/orders/0d9e8c7b-6a5f-4e3d-8c2b-1a0f9e8d7c6b/status
        "409":
          description: The requested transition is not allowed from the current order status.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              example:
                type: https://surplasse.com/problems/order-not-modifiable
                title: Order not modifiable
                status: 409
                detail: Order status cannot move from paid to ready.
                instance: /v1/orders/0d9e8c7b-6a5f-4e3d-8c2b-1a0f9e8d7c6b/status
        "422":
          description: The terminal status is incompatible with the order type.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              example:
                type: https://surplasse.com/problems/business-rule-violation
                title: Business rule violation
                status: 422
                detail: An on-site order must finish as served.
                instance: /v1/orders/0d9e8c7b-6a5f-4e3d-8c2b-1a0f9e8d7c6b/status
...
components:
  securitySchemes:
    tableSession:
      type: apiKey
      in: header
      name: X-Table-Session
      description: >
        Opaque anonymous table session token, delivered by
        `createTableSession` when a table QR is scanned. Bound to one
        establishment and one table, valid for the duration of a meal,
        carrying no personal data.
    restaurateurSession:
      type: apiKey
      in: cookie
      name: surplasse_session
      description: >
        Short session JWT set after the magic link exchange. The cookie is
        host-only for api.surplasse.com, HttpOnly, Secure in production,
        SameSite=Lax and scoped to Path=/. Used by Dashboard endpoints and SSE.
    restaurateurRefresh:
      type: apiKey
      in: cookie
      name: surplasse_refresh
      description: >
        Opaque rotating refresh token, hashed server side and valid for 30
        days. The host-only cookie is HttpOnly, Secure in production,
        SameSite=Lax and scoped to Path=/v1/auth/sessions.
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: >
        Client-generated UUID identifying this intention. Replaying the same
        request with the same key returns the original response without any
        duplicate; the same key with a different payload yields a 409
        `idempotency-key-conflict`.
      schema:
        type: string
        format: uuid
      example: 7f3a9c1e-4b2d-4e8a-9c5f-1d2e3f4a5b6c
    OrderId:
      name: orderId
      in: path
      required: true
      description: Identifier of the order.
      schema:
        type: string
        format: uuid
      example: 0d9e8c7b-6a5f-4e3d-8c2b-1a0f9e8d7c6b
    TrackingToken:
      name: trackingToken
      in: query
      required: true
      description: >
        Non-guessable tracking capability of the order, returned at
        creation. Carried by the tracking page URL so the customer can
        close the browser and come back.
      schema:
        type: string
        pattern: "^ot_[a-f0-9]{32}$"
      example: ot_5c3e1a9f7d2b4c8a6e0f2a4c6e8b0d1f
...
  schemas:
    Problem:
      type: object
      description: >
        RFC 9457 Problem Details document, the single error format of the
        API. The `type` URI is stable and identifies the applicative error.
      properties:
        type:
          type: string
          description: Stable URI identifying the applicative error type.
          default: about:blank
        title:
          type: string
          description: Short, human-readable summary of the error type.
        status:
          type: integer
          description: HTTP status code of this occurrence.
          minimum: 100
          maximum: 599
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence.
        instance:
          type: string
          description: URI reference of the request that produced the error.
      required: [title, status]
      example:
        type: https://surplasse.com/problems/resource-not-found
        title: Resource not found
        status: 404
        detail: No establishment matches slug 'le-cormoran'.
        instance: /v1/establishments/le-cormoran/menu
...
    TableSessionRequest:
      type: object
      description: The context carried by a scanned table QR code.
      properties:
        establishmentSlug:
          type: string
          description: Slug of the establishment, from the mini-site subdomain.
          pattern: "^[a-z0-9](?:[a-z0-9-]{0,61}[a-z0-9])?$"
        tableCode:
          type: string
          description: Non-guessable table code carried by the QR URL, never a sequential number.
          minLength: 1
      required: [establishmentSlug, tableCode]
      example:
        establishmentSlug: le-cormoran
        tableCode: tbl_2f8e6a4c0b9d7e1f
    TableSession:
      type: object
      description: >
        An anonymous table session. The token is opaque (a server-side
        reference, not a JWT) and carries no personal data.
      properties:
        token:
          type: string
          description: Opaque session token, sent back in the X-Table-Session header.
        establishmentId:
          type: string
          format: uuid
          description: Establishment the session is bound to.
        tableLabel:
          type: string
          description: Human label of the table, resolved server side from the scanned code.
        expiresAt:
          type: string
          format: date-time
          description: Expiry of the session, sliding while the session is active.
      required: [token, establishmentId, tableLabel, expiresAt]
      example:
        token: ts_9f2c4b7a1e6d8a3c5f0b2d4e6a8c1f3b
        establishmentId: 7c9e6679-7425-40de-944b-e07fc1f90ae7
        tableLabel: Table 4
        expiresAt: "2026-07-18T14:30:00Z"
    OrderCreationRequest:
      type: object
      description: >
        The validated cart sent by the Commande frontend. The establishment
        and the table come from the table session; amounts are recomputed
        server side from the catalog and never trusted from the client.
      properties:
        type:
          type: string
          description: Type of the order. Only `on_site` is accepted at this stage.
          enum: [on_site, takeaway]
        lines:
          type: array
          description: Lines of the cart, one per product and option combination.
          minItems: 1
          items:
            $ref: "#/components/schemas/OrderLineRequest"
      required: [type, lines]
      example:
        type: on_site
        lines:
          - productId: d0e1f2a3-b4c5-46d7-88e9-f0a1b2c3d405
            quantity: 1
            optionIds:
              - f0a1b2c3-d4e5-46f7-88a9-b0c1d2e3f404
            note: Sauce à part.
...
    Order:
      type: object
      description: >
        An order as seen by the customer: lines frozen at creation (names,
        prices and options copied from the catalog at that instant), status
        driven by the kitchen.
      properties:
        id:
          type: string
          format: uuid
          description: Identifier of the order.
        displayNumber:
          type: string
          description: Short number displayed to the customer and the staff, unique per establishment and day.
        status:
          type: string
          description: Current status of the order.
          enum:
            - pending_payment
            - paid
            - accepted
            - preparing
            - ready
            - served
            - picked_up
            - cancelled
            - refunded
        type:
          type: string
          description: Type of the order.
          enum: [on_site, takeaway]
        tableLabel:
          type: string
          description: Label of the table for on-site orders. Absent for takeaway.
        lines:
          type: array
          description: Lines of the order, frozen at creation.
          items:
            $ref: "#/components/schemas/OrderLine"
        totalCents:
          type: integer
          description: Total of the order in cents, sum of the line totals.
          minimum: 0
        currency:
          type: string
          description: ISO 4217 currency code.
        trackingToken:
          type: string
          description: Non-guessable capability giving access to the tracking page and stream.
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp, ISO 8601 UTC.
      required: [id, displayNumber, status, type, lines, totalCents, currency, trackingToken, createdAt]
      example:
        id: 0d9e8c7b-6a5f-4e3d-8c2b-1a0f9e8d7c6b
        displayNumber: "17"
        status: pending_payment
        type: on_site
        tableLabel: Table 4
        lines:
          - productId: d0e1f2a3-b4c5-46d7-88e9-f0a1b2c3d405
            productName: Burger du Vieux-Port
            unitPriceCents: 1600
            quantity: 1
            options:
              - group: Cuisson
                option: Saignant
                extraCostCents: 0
            note: Sauce à part.
            lineTotalCents: 1600
        totalCents: 1600
        currency: EUR
        trackingToken: ot_5c3e1a9f7d2b4c8a6e0f2a4c6e8b0d1f
        createdAt: "2026-07-18T12:34:56Z"
...
    OrderPage:
      type: object
      description: >
        One cursor-paginated page of operational orders, newest first.
        nextCursor is present exactly when hasMore is true.
      properties:
        items:
          type: array
          description: Operational orders in this page.
          items:
            $ref: "#/components/schemas/DashboardOrder"
        nextCursor:
          type: string
          minLength: 1
          maxLength: 512
          description: Opaque cursor to request the following page.
        hasMore:
          type: boolean
          description: Whether another page exists.
      required: [items, hasMore]
      example:
        items:
          - id: 0d9e8c7b-6a5f-4e3d-8c2b-1a0f9e8d7c6b
            displayNumber: "17"
            status: paid
            type: on_site
            tableLabel: Table 4
            lines: []
            totalCents: 1600
            currency: EUR
            createdAt: "2026-07-18T12:34:56Z"
        nextCursor: djF8MjAyNi0wNy0xOFQxMjozNDo1NlowZDllOGM3Yg
        hasMore: true
...